plotPairsArches innerHeight adjustment for arches with different sized anchors.
This commit addresses a bug referenced in issue #61.
The bezier curves for plotPairsArches would overlap when pair anchors were not
the same size (i.e. right anchor much larger than left anchor). Logic was
added to check if anchors are not the same size and then scale the inner arch height
based on the outer arch height. This logic was adjusted to work no matter
the orientation of the pair data input. For example,
`data.frame("chr10",26796593,26861030,"chr10",26735842,26739777)` and
`data.frame("chr10",26735842,26739777,"chr10",26796593,26861030)` should yield the same
plot.