Transcript highlighting, issue #54.
Functionality was added to plotTranscripts to highlight specific transcripts with colors
by either specifying them with their transcript name or their gene name. The gene
prioritization labeling logic in R/assemblyFunctions.R had to be slightly modified to allow
for highlights to be specified for transcripts. Associated documentation was added for
this feature as well as a new section in the "Aesthetics" vignette.