mapColors bug fix in reference to issue #63.
The mapColors function can now appropriately map colors to a numeric vector
that have all the same value, so long as a range is provided. This was
edited in the error checking of the mapColors code. Behavior will be as
before if no range is provided because function will not be able to determine
appropriate breaks.